Extract `FeedImageDataLoader` test helpers into a shared protocol extension to remove duplication
